+++ src/components/auth/signin/SigninFooter.tsx
@@ -3,11 +3,12 @@ import AuthFooterLayout from '../../../shared/util/ui/AuthFooterLayout';
 
 interface SigninFooterProps {
   totalValid : boolean;
+  onClick: () => void
 }
 
-const SigninFooter = ({totalValid}:SigninFooterProps) => {
+const SigninFooter = ({totalValid,onClick}:SigninFooterProps) => {
   return (
-    <AuthFooterLayout infoMessage="Eight 계정이 없으신가요?" linkUrl="/signup" linkMessage="클릭하여 계정을 생성하세요." buttonText="로그인" totalValid={totalValid} />
+    <AuthFooterLayout infoMessage="Eight 계정이 없으신가요?" linkUrl="/signup" linkMessage="클릭하여 계정을 생성하세요." buttonText="로그인" totalValid={totalValid} onClick={onClick} />
   );
 };
